+++ custom_components/hph/switch.py
@@ -15,7 +15,7 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.restore_state import RestoreEntity
 
-from .const import CTRL_FACADES, DOMAIN, INTEGRATION_NAME, SWITCH_HELPERS
+from .const import CTRL_FACADES, DOMAIN, SWITCH_HELPERS
 from .helpers.proxy import FacadeProxyMixin, call_domain_service
 
 
@@ -63,7 +63,6 @@ class HphSwitch(SwitchEntity, RestoreEntity):
     def device_info(self):
         return {
             "identifiers": {(DOMAIN, "hub")},
-            "name": INTEGRATION_NAME,
             "manufacturer": "HeatPump Hero",
             "model": "Bundle integration",
         }
@@ -118,7 +117,6 @@ class HphFacadeSwitch(FacadeProxyMixin, SwitchEntity):
     def device_info(self):
         return {
             "identifiers": {(DOMAIN, "hub")},
-            "name": INTEGRATION_NAME,
             "manufacturer": "HeatPump Hero",
             "model": "Bundle integration",
         }